homework

+++ unit_3/w12d03_lab_fibonacci/server.js
@@ -8,8 +8,8 @@ app.get('/', (req, res)=>{
 
 app.get('/:fibonacci_number', (req, res)=>{
     const {fibonacci_number} = req.params;
-    // const {fibonacci_number} = req.params;
-    const num = parseInt(fibonacci_number);
+    const num = parseFloat(fibonacci_number);
+
     function isSquare(n) {
         return Math.sqrt(n) % 1 === 0;
     }
@@ -25,3 +25,29 @@ app.get('/:fibonacci_number', (req, res)=>{
 app.listen(port, ()=>{
     console.log(`Listening on port ${port}`);
 })
+////////////////////////////////////////////////////
+// const isFibonacci = (num) => {
+//     if (num === 0 || num === 1) return true;
+
+//     let a = 0;
+//     let b = 1;
+//     while (b < num) {
+//         [a, b] = [b, a + b];
+//     }
+//     return b === num;
+// };
+
+
+///////////////////////////////////////////////////
+// app.get('/fibonacci/:number', (req, res) => {
+//     const num = parseInt(req.params.number);
+//     if (isNaN(num)) {
+//         res.send("Please provide a valid number.");
+//     } else {
+//         if (isFibonacci(num)) {
+//             res.send("Very good. It is Fibonacci.");
+//         } else {
+//             res.send("I can tell this is not a fibonacci number.");
+//         }
+//     }
+// });
